+++ lsm-tree/test/Test/Database/LSMTree/Internal/Snapshot/Codec.hs
@@ -39,21 +39,6 @@ tests = testGroup "Test.Database.LSMTree.Internal.Snapshot.Codec" [
         , testProperty "roundtripCBOR" $ roundtripCBOR (Proxy @SnapshotVersion)
         , testProperty "roundtripFlatTerm" $ roundtripFlatTerm (Proxy @SnapshotVersion)
         ]
-    , testGroup "snapshot version errors" [
-          testProperty "unknown versions are rejected" $ \(NonNegative n) ->
-            let w = 3 + fromIntegral (n :: Int) in
-            case snapshotVersionFromWord w of
-              Left (ErrSnapshotVersionUnknown found current) ->
-                found === w .&&. current === currentSnapshotVersion
-              Right v -> counterexample ("unexpectedly known: " <> show v) False
-        , testProperty "all compatible versions pass checkCompatible" $
-            forAll (elements allCompatibleSnapshotVersions) $ \v ->
-              checkCompatible v === Right ()
-        , testProperty "decodeVersionHeader agrees with encodeSnapshotMetaData" $ \metadata ->
-            case decodeVersionHeader (encodeSnapshotMetaData metadata) of
-              Left err     -> counterexample err False
-              Right (_, w) -> snapshotVersionFromWord w === Right currentSnapshotVersion
-        ]
     , testGroup "Versioned SnapshotMetaData" [
           testProperty "roundtripCBOR" $ roundtripCBOR (Proxy @(Versioned SnapshotMetaData))
         , testProperty "roundtripFlatTerm" $ roundtripFlatTerm (Proxy @(Versioned SnapshotMetaData))
@@ -75,9 +60,9 @@ tests = testGroup "Test.Database.LSMTree.Internal.Snapshot.Codec" [
 
 -- | @decode . encode = id@
 explicitRoundtripCBOR ::
-     (Eq a, Show a)
+     (Eq a, Show a, Eq e, Show e)
   => (a -> Encoding)
-  -> (forall s. Decoder s a)
+  -> (forall s. Decoder s (Either e a))
   -> a
   -> Property
 explicitRoundtripCBOR enc dec x = case back (there x) of
@@ -88,7 +73,7 @@ explicitRoundtripCBOR enc dec x = case back (there x) of
     Right (bs, y) ->
            counterexample
              ("Expected value and roundtripped value do not match")
-             (x === y)
+             (Right x === y)
       .&&. counterexample
              ("Found trailing bytes")
              (property (BSL.null bs))
@@ -102,15 +87,15 @@ roundtripCBOR _ = explicitRoundtripCBOR encode decode
 
 -- | See 'explicitRoundtripCBOR'.
 roundtripCBOR' :: (Encode a, DecodeVersioned a, Eq a, Show a) => Proxy a -> a -> Property
-roundtripCBOR' _ = explicitRoundtripCBOR encode (decodeVersioned currentSnapshotVersion)
+roundtripCBOR' _ = explicitRoundtripCBOR encode (Right @() <$> decodeVersioned currentSnapshotVersion)
 
 -- | @fromFlatTerm . toFlatTerm = id@
 --
 -- This will also check @validFlatTerm@ on the result of @toFlatTerm@.
 explicitRoundtripFlatTerm ::
-     (Eq a, Show a)
+     (Eq a, Show a, Eq e, Show e)
   => (a -> Encoding)
-  -> (forall s. Decoder s a)
+  -> (forall s. Decoder s (Either e a))
   -> a
   -> Property
 explicitRoundtripFlatTerm enc dec x = case back flatTerm of
@@ -121,7 +106,7 @@ explicitRoundtripFlatTerm enc dec x = case back flatTerm of
     Right y ->
            counterexample
              ("Expected value and roundtripped value do not match")
-             (x === y)
+             (Right x === y)
       .&&. counterexample
              ("Invalid flat term")
              (property (validFlatTerm flatTerm))
@@ -145,7 +130,7 @@ roundtripFlatTerm' ::
   => Proxy a
   -> a
   -> Property
-roundtripFlatTerm' _ = explicitRoundtripFlatTerm encode (decodeVersioned currentSnapshotVersion)
+roundtripFlatTerm' _ = explicitRoundtripFlatTerm encode (Right @() <$> decodeVersioned currentSnapshotVersion)
 
 {-------------------------------------------------------------------------------
   Test and property runners